test(azure): isolate deployment path safety coverage
